+++ examples/platformer/climbing.rs
@@ -1,0 +1,82 @@
+use bevy::{prelude::*, utils::HashSet};
+use bevy_ecs_ldtk::prelude::*;
+use bevy_rapier2d::prelude::*;
+
+use crate::physics::SensorBundle;
+
+#[derive(Clone, Eq, PartialEq, Debug, Default, Component)]
+pub struct Climber {
+    pub climbing: bool,
+    pub intersecting_climbables: HashSet<Entity>,
+}
+
+#[derive(Copy, Clone, Eq, PartialEq, Debug, Default, Component)]
+pub struct Climbable;
+
+#[derive(Clone, Default, Bundle, LdtkIntCell)]
+pub struct LadderBundle {
+    #[from_int_grid_cell]
+    pub sensor_bundle: SensorBundle,
+    pub climbable: Climbable,
+}
+
+
+pub fn detect_climb_range(
+    mut climbers: Query<&mut Climber>,
+    climbables: Query<Entity, With<Climbable>>,
+    mut collisions: EventReader<CollisionEvent>,
+) {
+    for collision in collisions.read() {
+        match collision {
+            CollisionEvent::Started(collider_a, collider_b, _) => {
+                if let (Ok(mut climber), Ok(climbable)) =
+                    (climbers.get_mut(*collider_a), climbables.get(*collider_b))
+                {
+                    climber.intersecting_climbables.insert(climbable);
+                }
+                if let (Ok(mut climber), Ok(climbable)) =
+                    (climbers.get_mut(*collider_b), climbables.get(*collider_a))
+                {
+                    climber.intersecting_climbables.insert(climbable);
+                };
+            }
+            CollisionEvent::Stopped(collider_a, collider_b, _) => {
+                if let (Ok(mut climber), Ok(climbable)) =
+                    (climbers.get_mut(*collider_a), climbables.get(*collider_b))
+                {
+                    climber.intersecting_climbables.remove(&climbable);
+                }
+
+                if let (Ok(mut climber), Ok(climbable)) =
+                    (climbers.get_mut(*collider_b), climbables.get(*collider_a))
+                {
+                    climber.intersecting_climbables.remove(&climbable);
+                }
+            }
+        }
+    }
+}
+
+pub fn ignore_gravity_if_climbing(
+    mut query: Query<(&Climber, &mut GravityScale), Changed<Climber>>,
+) {
+    for (climber, mut gravity_scale) in &mut query {
+        if climber.climbing {
+            gravity_scale.0 = 0.0;
+        } else {
+            gravity_scale.0 = 1.0;
+        }
+    }
+}
+
+
+pub struct ClimbingPlugin;
+
+impl Plugin for ClimbingPlugin {
+	fn build(&self, app: &mut App) {
+		app
+        .add_systems(Update, detect_climb_range)
+        .add_systems(Update, ignore_gravity_if_climbing)
+		;
+	}
+}
