--- examples/platformer/components.rs
@@ -1,233 +1,0 @@
-use bevy::prelude::*;
-use bevy_ecs_ldtk::{prelude::*, utils::ldtk_pixel_coords_to_translation_pivoted};
-
-use std::collections::HashSet;
-
-use bevy_rapier2d::prelude::*;
-
-#[derive(Clone, Default, Bundle, LdtkIntCell)]
-pub struct ColliderBundle {
-    pub collider: Collider,
-    pub rigid_body: RigidBody,
-    pub velocity: Velocity,
-    pub rotation_constraints: LockedAxes,
-    pub gravity_scale: GravityScale,
-    pub friction: Friction,
-    pub density: ColliderMassProperties,
-}
-
-#[derive(Clone, Default, Bundle, LdtkIntCell)]
-pub struct SensorBundle {
-    pub collider: Collider,
-    pub sensor: Sensor,
-    pub active_events: ActiveEvents,
-    pub rotation_constraints: LockedAxes,
-}
-
-impl From<&EntityInstance> for ColliderBundle {
-    fn from(entity_instance: &EntityInstance) -> ColliderBundle {
-        let rotation_constraints = LockedAxes::ROTATION_LOCKED;
-
-        match entity_instance.identifier.as_ref() {
-            "Player" => ColliderBundle {
-                collider: Collider::cuboid(6., 14.),
-                rigid_body: RigidBody::Dynamic,
-                friction: Friction {
-                    coefficient: 0.0,
-                    combine_rule: CoefficientCombineRule::Min,
-                },
-                rotation_constraints,
-                ..Default::default()
-            },
-            "Mob" => ColliderBundle {
-                collider: Collider::cuboid(5., 5.),
-                rigid_body: RigidBody::KinematicVelocityBased,
-                rotation_constraints,
-                ..Default::default()
-            },
-            "Chest" => ColliderBundle {
-                collider: Collider::cuboid(8., 8.),
-                rigid_body: RigidBody::Dynamic,
-                rotation_constraints,
-                gravity_scale: GravityScale(1.0),
-                friction: Friction::new(0.5),
-                density: ColliderMassProperties::Density(15.0),
-                ..Default::default()
-            },
-            _ => ColliderBundle::default(),
-        }
-    }
-}
-
-impl From<IntGridCell> for SensorBundle {
-    fn from(int_grid_cell: IntGridCell) -> SensorBundle {
-        let rotation_constraints = LockedAxes::ROTATION_LOCKED;
-
-        // ladder
-        if int_grid_cell.value == 2 {
-            SensorBundle {
-                collider: Collider::cuboid(8., 8.),
-                sensor: Sensor,
-                rotation_constraints,
-                active_events: ActiveEvents::COLLISION_EVENTS,
-            }
-        } else {
-            SensorBundle::default()
-        }
-    }
-}
-
-#[derive(Clone, Component, Debug, Eq, Default, PartialEq)]
-pub struct Items(Vec<String>);
-
-impl From<&EntityInstance> for Items {
-    fn from(entity_instance: &EntityInstance) -> Self {
-        Items(
-            entity_instance
-                .iter_enums_field("items")
-                .expect("items field should be correctly typed")
-                .cloned()
-                .collect(),
-        )
-    }
-}
-
-#[derive(Copy, Clone, Eq, PartialEq, Debug, Default, Component)]
-pub struct Player;
-
-#[derive(Clone, Eq, PartialEq, Debug, Default, Component)]
-pub struct Climber {
-    pub climbing: bool,
-    pub intersecting_climbables: HashSet<Entity>,
-}
-
-#[derive(Clone, Default, Bundle, LdtkEntity)]
-pub struct PlayerBundle {
-    #[sprite_bundle("player.png")]
-    pub sprite_bundle: SpriteBundle,
-    #[from_entity_instance]
-    pub collider_bundle: ColliderBundle,
-    pub player: Player,
-    #[worldly]
-    pub worldly: Worldly,
-    pub climber: Climber,
-    pub ground_detection: GroundDetection,
-
-    // Build Items Component manually by using `impl From<&EntityInstance>`
-    #[from_entity_instance]
-    items: Items,
-
-    // The whole EntityInstance can be stored directly as an EntityInstance component
-    #[from_entity_instance]
-    entity_instance: EntityInstance,
-}
-
-#[derive(Copy, Clone, Eq, PartialEq, Debug, Default, Component)]
-pub struct Wall;
-
-#[derive(Clone, Debug, Default, Bundle, LdtkIntCell)]
-pub struct WallBundle {
-    wall: Wall,
-}
-
-#[derive(Copy, Clone, Eq, PartialEq, Debug, Default, Component)]
-pub struct Climbable;
-
-#[derive(Clone, Default, Bundle, LdtkIntCell)]
-pub struct LadderBundle {
-    #[from_int_grid_cell]
-    pub sensor_bundle: SensorBundle,
-    pub climbable: Climbable,
-}
-
-#[derive(Copy, Clone, Eq, PartialEq, Debug, Default, Component)]
-pub struct Enemy;
-
-#[derive(Clone, PartialEq, Debug, Default, Component)]
-pub struct Patrol {
-    pub points: Vec<Vec2>,
-    pub index: usize,
-    pub forward: bool,
-}
-
-impl LdtkEntity for Patrol {
-    fn bundle_entity(
-        entity_instance: &EntityInstance,
-        layer_instance: &LayerInstance,
-        _: Option<&Handle<Image>>,
-        _: Option<&TilesetDefinition>,
-        _: &AssetServer,
-        _: &mut Assets<TextureAtlasLayout>,
-    ) -> Patrol {
-        let mut points = Vec::new();
-        points.push(ldtk_pixel_coords_to_translation_pivoted(
-            entity_instance.px,
-            layer_instance.c_hei * layer_instance.grid_size,
-            IVec2::new(entity_instance.width, entity_instance.height),
-            entity_instance.pivot,
-        ));
-
-        let ldtk_patrol_points = entity_instance
-            .iter_points_field("patrol")
-            .expect("patrol field should be correclty typed");
-
-        for ldtk_point in ldtk_patrol_points {
-            // The +1 is necessary here due to the pivot of the entities in the sample
-            // file.
-            // The patrols set up in the file look flat and grounded,
-            // but technically they're not if you consider the pivot,
-            // which is at the bottom-center for the skulls.
-            let pixel_coords = (ldtk_point.as_vec2() + Vec2::new(0.5, 1.))
-                * Vec2::splat(layer_instance.grid_size as f32);
-
-            points.push(ldtk_pixel_coords_to_translation_pivoted(
-                pixel_coords.as_ivec2(),
-                layer_instance.c_hei * layer_instance.grid_size,
-                IVec2::new(entity_instance.width, entity_instance.height),
-                entity_instance.pivot,
-            ));
-        }
-
-        Patrol {
-            points,
-            index: 1,
-            forward: true,
-        }
-    }
-}
-
-#[derive(Clone, Default, Bundle, LdtkEntity)]
-pub struct MobBundle {
-    #[sprite_sheet_bundle]
-    pub sprite_sheet_bundle: LdtkSpriteSheetBundle,
-    #[from_entity_instance]
-    pub collider_bundle: ColliderBundle,
-    pub enemy: Enemy,
-    #[ldtk_entity]
-    pub patrol: Patrol,
-}
-
-#[derive(Clone, Default, Bundle, LdtkEntity)]
-pub struct ChestBundle {
-    #[sprite_sheet_bundle]
-    pub sprite_sheet_bundle: LdtkSpriteSheetBundle,
-    #[from_entity_instance]
-    pub collider_bundle: ColliderBundle,
-}
-
-#[derive(Clone, Default, Bundle, LdtkEntity)]
-pub struct PumpkinsBundle {
-    #[sprite_sheet_bundle(no_grid)]
-    pub sprite_sheet_bundle: LdtkSpriteSheetBundle,
-}
-
-#[derive(Clone, Default, Component)]
-pub struct GroundDetection {
-    pub on_ground: bool,
-}
-
-#[derive(Component)]
-pub struct GroundSensor {
-    pub ground_detection_entity: Entity,
-    pub intersecting_ground_entities: HashSet<Entity>,
-}
